requirements.txt
aws-cdk-lib>=2.127.0
constructs>=10.3.0
boto3>=1.34
requests==2.31.0
black==24.1
flake8>=7
isort>=5.13